Split code into modules based on functionality
